backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
python-dateutil==2.8.2
requests==2.31.0
python-dotenv==1.0.0
